requirements.txt
pytrends
requests
selenium
nomics-python
python-binance